Only show the Open in Volview button for some file extensions.
This will only show the Open in Volview button for item lists if at
least one item has a final extension that matches a list.  For item
pages, the button will only show if the item has such an extension.

This is still imperfect -- item extensions are not required, and the
extensions are the wrong granularity (e.g., .gz is too inclusive).  This
is still an improvement over the current always-show-button behavior.

In the future, we can tune the extensions or add mime-type detection as
well (or even check some other configuration).