Use mongo aggregations when checking if a folder can be opened
I noticed that the check for if a folder should be opened by volview is
very slow.  This is entirely the fault of the iterator that is used in
Girder to get files from a folder -- it recurses subfolders and items
with function calls, each of which results in a mongo query.  It also
can extract metadata as if it were a file (the function supports the
download of folders, and wasn't optimized for just listing files by
themselves).  In my test system I have some directories with a lot of
items (30,000+), and the volview for check takes 45+ seconds.

Switching to to a mongo aggregation uses a single database roundtrip
(instead, in my example of 30,000+ database roundtrips), dropping the
time to a few hundred milliseconds.

It might also be worth caching the check of a folder for some duration.