Better open volview from item lists
This requires a new enough version of large_image to be useful.

Specifically, this exposes open buttons on items in item lists.  If the
item lists are grouped and filtered, this will open the filter set and
have appropriately named session files.  Filters have a pipeline that
given a parent folder find all the items that match a filter in any
recursion level and then add the files of those items to the open set.
This is passed compactly, so any number of files can be opened with
short paths.

Note that this does nothing to change the existing checked item and
folder behavior.  A future work would be to generalize folder-level
"app" open behavior.  This will require determining whether a list of
folders can each be opened, which is currently an expensive call.

Although there is a tox lint environment, it doesn't seem to be enforced
in any manner.  I'd recommend adding an autoformater (either autopep8 or
black, etc.) for both python and javascript.