Ensure plugin load order
The client requires large_image.  It seems like it _should_ be
sufficient to say this in the package.json, but the order the javascript
is loaded is dictated by the python order.  Without asking to load
large_image, it is possible the girder volview plugin will be loaded
first and fail.